com.feilong.io.IOWriteUtil.write(InputStream, OutputStream) 2.57G的文件 日志 FileUtil.formatSize(long) 会报错 fix #528

+++ feilong-io/src/main/java/com/feilong/io/IOWriteUtil.java
@@ -421,7 +421,7 @@ public final class IOWriteUtil{
         Date beginDate = now();
 
         int loopCount = 0;
-        int sumSize = 0;
+        long sumSize = 0;
         ByteBuffer byteBuffer = ByteBuffer.allocate(bufferLength);
 
         try (ReadableByteChannel readableByteChannel = Channels.newChannel(inputStream);
